Fix ROS builds
This packge is not a catkin package - there is no mention of catkin in
the CMakeLists.txt. Catkin can still build the package if it is a 'plain
cmake' package.

Also, the Cairo dependency was missing.

The package.xml should really be installed to the share directory as
well, so that when this package is installed, rosdep can see it.